Add files via upload
This project was conducted to satisfy the requirements of the "Mini-Project" module in the 4th Semester of our Bachelor of Science Engineering (Hons.) in Mechatronics Engineering Degree Program at the General Sir John Kotelawala Defence University in Ratmalana, Sri Lanka.
The fabricated device applies an Image Processing Algorithm on each frame of a video stream being captured by a Raspberry Pi V2 Camera to track the location and velocity of a ping pong ball on an acrylic plate, and after a PID calculation, adjusts the plate's orientation using Parallel Manipulators actuated by MG996R Servo Motors to center the ball. A Raspberry Pi 4 Model B was used as the microprocessor, as well as a PCA9685 16 Channel Servo Motor Driver to control the Servo Motors.
The project won the title of "Best Project".